requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
python-multipart==0.0.9
llama-cpp-python==0.2.55
pyinstaller==6.3.0
